# TarotReading/specs/001-tarot-reading/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: AI Tarot Friend API
  description: REST API for empathetic tarot reading platform with memory and proactive care
  version: 1.0.0
  contact:
    name: AI Tarot Friend Team
servers:
  - url: https://api.aitarotfriend.com/v1
    description: Production
  - url: https://staging-api.aitarotfriend.com/v1
    description: Staging
  - url: http://localhost:3000/v1
    description: Local Development
security:
  - BearerAuth: []
paths:
...
  /readings:
    post:
      summary: Create a new tarot reading
      description: Draw cards for a spread and generate interpretation
      tags: [Readings]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [spread_type]
              properties:
                spread_type:
                  type: string
                  enum: ['1-card', '3-card', '7-card', 'celtic-cross']
                seed:
                  type: string
                  description: Optional seed for reproducibility (premium only)
                context:
                  type: string
                  description: User's question or concern
      responses:
        '201':
          description: Reading created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadingResponse'
        '429':
          $ref: '#/components/responses/QuotaExceeded'
        '403':
          $ref: '#/components/responses/PremiumRequired'

    get:
      summary: Get reading history
      description: List past readings for authenticated user
      tags: [Readings]
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Reading list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Reading'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
  /readings/{reading_id}:
    get:
      summary: Get specific reading details
      tags: [Readings]
      parameters:
        - name: reading_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Reading details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reading'
        '404':
          $ref: '#/components/responses/NotFound'

  # Session Endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ReadingResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        spread_type:
          type: string
        cards:
          type: array
          items:
            $ref: '#/components/schemas/DrawnCard'
        interpretation:
          $ref: '#/components/schemas/Interpretation'
        created_at:
          type: string
          format: date-time
    Reading:
      allOf:
        - $ref: '#/components/schemas/ReadingResponse'
        - type: object
          properties:
            session_id:
              type: string
              format: uuid
            token_count:
              type: integer
    DrawnCard:
      type: object
      properties:
        card_id:
          type: integer
        name:
          type: string
        position:
          type: integer
        orientation:
          type: string
          enum: [upright, reversed]
        meaning:
          type: string
    Interpretation:
      type: object
      properties:
        tldr:
          type: string
        key_points:
          type: array
          items:
            type: string
          minItems: 3
          maxItems: 5
        advice:
          type: object
          properties:
            short_term:
              type: string
            medium_term:
              type: string
            long_term:
              type: string
        warnings:
          type: string
...
    Pagination:
      type: object
      properties:
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
  responses:
    QuotaExceeded:
      description: Daily quota exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: QUOTA_EXCEEDED
              message: You've reached your daily limit. Upgrade to Premium for unlimited readings.
    PremiumRequired:
      description: Premium subscription required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: PREMIUM_REQUIRED
              message: This spread type is available for Premium subscribers only.
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'